MES-257 - Database/Entity: add counting_equipment columns/properties

+++ src/main/java/com/tde/mescloud/model/entity/CountingEquipmentEntity.java
@@ -30,7 +30,7 @@ public class CountingEquipmentEntity {
     List<ProductionOrderEntity> productionOrders;
 
     private double equipmentEffectiveness;
-    private double theoreticalProduction;
+    private int theoreticalProduction;
     private double availability;
     private double performance;
     private double quality;
